# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Indexer APIs for Andamio dapp-system.",
        "title": "Andamio Indexer 1.0.0",
        "contact": {
            "name": "Andamio Support",
            "url": "https://www.andamio.com/support",
            "email": "dev@andamio.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "host": "0.0.0.0:42069",
    "basePath": "/api/v1/indexer",
    "paths": {
...
        "/assets/{asset_fingerprint}/addresses": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Retrieve a list of addresses that hold a specific asset fingerprint, with support for pagination.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Assets"
                ],
                "summary": "Get Addresses by Asset Fingerprint",
                "operationId": "getAddressesByAssetFingerprint",
                "parameters": [
                    {
                        "type": "string",
                        "description": "The asset fingerprint (hex-encoded) to retrieve addresses for.",
                        "name": "asset_fingerprint",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "default": 100,
                        "description": "Maximum number of results to return.",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 0,
                        "description": "Number of results to skip.",
                        "name": "offset",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully retrieved addresses.",
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid asset fingerprint or pagination parameters.",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "error": {
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Asset fingerprint not found or no addresses found.",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "error": {
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error.",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "error": {
                                    "type": "string"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/indexer/assets/fingerprint/{asset_fingerprint}/transactions": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Retrieves a list of transactions associated with a specific asset fingerprint, with support for pagination.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Assets"
                ],
                "summary": "Get Transactions by Asset Fingerprint",
                "operationId": "getTransactionsByAssetFingerprint",
                "parameters": [
                    {
                        "type": "string",
                        "description": "The asset fingerprint (hex-encoded) to retrieve transactions for.",
                        "name": "asset_fingerprint",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "default": 100,
                        "description": "Maximum number of results to return.",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 0,
                        "description": "Number of results to skip.",
                        "name": "offset",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully retrieved transactions.",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/viewmodel.Transaction"
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid asset fingerprint or pagination parameters.",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "error": {
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Asset fingerprint not found or no transactions found.",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "error": {
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error.",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "error": {
                                    "type": "string"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/indexer/assets/policy/{policyId}/token/{tokenname}/transactions": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Retrieves transactions associated with a given policy ID and token name with pagination.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Assets"
                ],
                "summary": "Get Transactions by Policy ID and Token Name",
                "operationId": "getTransactionsByPolicyIdAndTokenName",
                "parameters": [
                    {
                        "type": "string",
                        "description": "The policy ID to retrieve transactions for (hex-encoded).",
                        "name": "policyId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "The token name to retrieve transactions for (hex-encoded).",
                        "name": "tokenname",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "default": 100,
                        "description": "Maximum number of results to return.",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 0,
                        "description": "Number of results to skip.",
                        "name": "offset",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully retrieved transactions.",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/viewmodel.Transaction"
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid policy ID, token name, or pagination parameters.",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "error": {
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Policy ID and token name combination not found or no transactions found.",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "error": {
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error.",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "error": {
                                    "type": "string"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "viewmodel.Asset": {
            "type": "object",
            "properties": {
                "amount": {
                    "type": "integer"
                },
                "fingerprint": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "name_hex": {
                    "type": "string"
                },
                "policy_id": {
                    "type": "string"
                },
                "utxo_id": {
                    "type": "string"
                },
                "utxo_index": {
                    "type": "integer"
                }
            }
        },
        "viewmodel.Datum": {
            "type": "object",
            "properties": {
                "datum_cbor": {
                    "type": "string"
                },
                "datum_hash": {
                    "type": "string"
                },
                "utxo_id": {
                    "type": "string"
                },
                "utxo_index": {
                    "type": "integer"
                }
            }
        },
        "viewmodel.Redeemer": {
            "type": "object",
            "properties": {
                "cbor": {
                    "description": "CBOR string representation",
                    "type": "string"
                },
                "index": {
                    "type": "integer"
                },
                "tag": {
                    "type": "integer"
                },
                "transaction_hash": {
                    "type": "string"
                }
            }
        },
        "viewmodel.SimpleUTxO": {
            "type": "object",
            "properties": {
                "transaction_hash": {
                    "type": "string"
                },
                "utxo_id": {
                    "type": "string"
                },
                "utxo_index": {
                    "type": "integer"
                }
            }
        },
        "viewmodel.Transaction": {
            "type": "object",
            "properties": {
                "block_hash": {
                    "type": "string"
                },
                "block_number": {
                    "type": "integer"
                },
                "certificates": {
                    "description": "Slice of CBOR string representations",
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "fee": {
                    "type": "integer"
                },
                "inputs": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/viewmodel.TransactionInput"
                    }
                },
                "metadata": {
                    "description": "CBOR string representation",
                    "type": "string"
                },
                "outputs": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/viewmodel.TransactionOutput"
                    }
                },
                "reference_inputs": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/viewmodel.SimpleUTxO"
                    }
                },
                "slot_number": {
                    "type": "integer"
                },
                "transaction_hash": {
                    "type": "string"
                },
                "ttl": {
                    "type": "integer"
                },
                "withdrawals": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "integer"
                    }
                },
                "witness": {
                    "$ref": "#/definitions/viewmodel.Witness"
                }
            }
        },
        "viewmodel.TransactionInput": {
            "type": "object",
            "properties": {
                "address": {
                    "type": "string"
                },
                "amount": {
                    "type": "integer"
                },
                "asset": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/viewmodel.Asset"
                    }
                },
                "cbor": {
                    "description": "CBOR string representation",
                    "type": "string"
                },
                "datum": {
                    "$ref": "#/definitions/viewmodel.Datum"
                },
                "transaction_hash": {
                    "type": "string"
                },
                "utxo_id": {
                    "type": "string"
                },
                "utxo_index": {
                    "type": "integer"
                }
            }
        },
        "viewmodel.TransactionOutput": {
            "type": "object",
            "properties": {
                "address": {
                    "type": "string"
                },
                "amount": {
                    "type": "integer"
                },
                "asset": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/viewmodel.Asset"
                    }
                },
                "cbor": {
                    "description": "CBOR string representation",
                    "type": "string"
                },
                "datum": {
                    "$ref": "#/definitions/viewmodel.Datum"
                },
                "utxo_id": {
                    "type": "string"
                },
                "utxo_index": {
                    "type": "integer"
                }
            }
        },
...
        "viewmodel.Witness": {
            "type": "object",
            "properties": {
                "plutus_data": {
                    "description": "Slice of CBOR string representations",
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "plutus_v1_scripts": {
                    "description": "Slice of CBOR string representations",
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "plutus_v2_scripts": {
                    "description": "Slice of CBOR string representations",
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "plutus_v3_scripts": {
                    "description": "Slice of CBOR string representations",
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "redeemers": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/viewmodel.Redeemer"
                    }
                },
                "transaction_hash": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "BasicAuth": {
            "type": "basic"
        }
    },
...
}